requirements.txt
pandas==2.3.3
numpy==2.4.3
python-dotenv==1.2.2
sqlalchemy==2.0.48
matplotlib==3.10.8
matplotlib-inline==0.2.1
pyodbc==5.3.0
plotly==5.24.1
nbformat>=4.2.0
